Add appropriate permissions to our /etc structure
This, combined with the fingerprint, should fix all of the market issues
for HTC Hero.  Users should be able to see copy protected applications
as well as other 'one-off' applications that were previously missing.
Users will still need to use MarketEnabled in order to view paid
applications in markets where this is not allowed.