Update the list of packages, and add hdpi locale
Adding hdpi in the locales is required in order to ensure AOSP
applications that have hdpi resources, but not mdpi resources actually
have those files included (Gallery3D, for example).   The device can
scale down the images, but only if they exist and when building they are
only included for the display types specified.